improved task inputs
Gradle sometimes claims that all tasks in the project are up-to-date while a rebuild would be required.

When modifying a model file (*.mps) only the assembleMps doesn't recognize a changed input and gradle considers it up-to-date.
Added an input files dependency to every *.mps file in the project directory.

The same applies for changes to a module file that doesn't cause a change in build model. While these changes are very rare they sometimes happen and could cause strange errors at runtime. In my case I was getting class not found exception because a jar was missing in the build.xml libraries of the generate task.